wind/requirements.txt
# Core
fastapi==0.95.0
uvicorn==0.21.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
pydantic[email]==1.10.7

# Database
sqlalchemy==2.0.10
alembic==1.10.3
psycopg2-binary==2.9.5
redis==4.5.5
elasticsearch==8.7.0
sqlalchemy-utils==1.0.0

# Kafka
kafka-python==2.0.2
confluent-kafka==2.1.1

# ML
scikit-learn==1.2.2
pandas==2.0.1
numpy==1.24.3
joblib==1.2.0

# Network Monitoring
scapy==2.5.0
celery==5.3.1
flower==2.0.1

# Utils
python-dateutil==2.8.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Testing
pytest==7.3.1
httpx==0.24.0
pytest-asyncio==0.21.0
